Set !{tenantTo} instead of !{location.state.tenantTo} in holdings/items manifest
