+++ src/ViewHoldingsRecord.js
@@ -81,14 +81,7 @@ class ViewHoldingsRecord extends React.Component {
       path: 'holdings-storage/holdings/:{holdingsrecordid}',
       resourceShouldRefresh: false,
       accumulate: true,
-      tenant: (_q, _p, _r, _l, props) => {
-        const {
-          stripes,
-          location,
-        } = props;
-
-        return location?.state?.tenantTo || stripes.okapi.tenant;
-      },
+      tenant: '!{tenantTo}',
     },
     items: {
       type: 'okapi',
@@ -104,14 +97,7 @@ class ViewHoldingsRecord extends React.Component {
       type: 'okapi',
       path: 'inventory/instances/:{id}',
       accumulate: true,
-      tenant: (_q, _p, _r, _l, props) => {
-        const {
-          stripes,
-          location,
-        } = props;
-
-        return location?.state?.tenantTo || stripes.okapi.tenant;
-      },
+      tenant: '!{tenantTo}',
     },
     tagSettings: {
       type: 'okapi',
