+++ src/routes/ItemRoute.js
@@ -1,204 +1,51 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  flowRight,
-  get,
-} from 'lodash';
+import { flowRight } from 'lodash';
 
 import { stripesConnect } from '@folio/stripes/core';
 
-import { requestsStatusString } from '../Instance/ViewRequests/utils';
-
 import withLocation from '../withLocation';
 import { ItemView } from '../views';
-import { PaneLoading } from '../components';
 import { DataContext } from '../contexts';
-
-const getRequestsPath = `circulation/requests?query=(itemId==:{itemid}) and status==(${requestsStatusString}) sortby requestDate desc&limit=1`;
+import { switchAffiliation } from '../utils';
 
 class ItemRoute extends React.Component {
-  static manifest = Object.freeze({
-    query: {},
-    itemsResource: {
-      type: 'okapi',
-      path: 'inventory/items/:{itemid}',
-      POST: { path: 'inventory/items' },
-      resourceShouldRefresh: true,
-      tenant: (_q, _p, _r, _l, props) => {
-        const {
-          stripes,
-          location,
-        } = props;
-
-        return location?.state?.tenantTo || stripes.okapi.tenant;
-      },
-    },
-    markItemAsWithdrawn: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-withdrawn',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markItemAsMissing: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-missing',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsInProcess: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-in-process',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsInProcessNonRequestable: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-in-process-non-requestable',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsIntellectualItem: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-intellectual-item',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsLongMissing: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-long-missing',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsRestricted: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-restricted',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsUnavailable: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-unavailable',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    markAsUnknown: {
-      type: 'okapi',
-      POST: {
-        path: 'inventory/items/:{itemid}/mark-unknown',
-      },
-      clientGeneratePk: false,
-      fetch: false,
-    },
-    holdingsRecords: {
-      type: 'okapi',
-      path: 'holdings-storage/holdings/:{holdingsrecordid}',
-      tenant: (_q, _p, _r, _l, props) => {
-        const {
-          stripes,
-          location,
-        } = props;
-
-        return location?.state?.tenantTo || stripes.okapi.tenant;
-      },
-    },
-    instanceRecords: {
-      type: 'okapi',
-      path: 'inventory/instances/:{id}',
-      resourceShouldRefresh: true,
-    },
-    servicePoints: {
-      type: 'okapi',
-      path: 'service-points',
-      records: 'servicepoints',
-      params: (_q, _p, _r, _l, props) => {
-        // Only one service point is of interest here: the SP used for the item's last check-in
-        // (if the item has a check-in). Iff that service point ID is found, add a query param
-        // to filter down to that one service point in the records returned.
-        const servicePointId = get(props.resources, 'itemsResource.records[0].lastCheckIn.servicePointId', '');
-        const query = servicePointId && `id==${servicePointId}`;
-        return query ? { query } : {};
-      },
-      resourceShouldRefresh: true,
-    },
-    staffMembers: {
-      type: 'okapi',
-      path: 'users',
-      records: 'users',
-      params: (_q, _p, _r, _l, props) => {
-        const staffMemberId = get(props.resources, 'itemsResource.records[0].lastCheckIn.staffMemberId', '');
-        const query = staffMemberId && `id==${staffMemberId}`;
-
-        return query ? { query } : null;
-      },
-      resourceShouldRefresh: true,
-    },
-    // return a count of the requests matching the given item and status
-    requests: {
-      type: 'okapi',
-      path: getRequestsPath,
-      records: 'requests',
-      PUT: { path: 'circulation/requests/%{requestOnItem.id}' },
-    },
-    openLoans: {
-      type: 'okapi',
-      path: 'circulation/loans',
-      params: {
-        query: 'status.name=="Open" and itemId==:{itemid}',
-      },
-      records: 'loans',
-    },
-    requestOnItem: {},
-    tagSettings: {
-      type: 'okapi',
-      records: 'configs',
-      path: 'configurations/entries?query=(module==TAGS and configName==tags_enabled)',
-    },
-  });
-
-  isLoading = () => {
+  goBack = () => {
     const {
-      resources: {
-        itemsResource,
-        holdingsRecords,
-        instanceRecords,
-      },
+      match: { params: { id } },
+      location: { search },
+      history,
     } = this.props;
 
-    if (!itemsResource?.hasLoaded ||
-      !instanceRecords?.hasLoaded ||
-      !holdingsRecords?.hasLoaded) {
-      return true;
-    }
+    history.push({
+      pathname: `/inventory/view/${id}`,
+      search,
+    });
+  }
 
-    return false;
+  onClose = () => {
+    const {
+      stripes,
+      location,
+    } = this.props;
+    const tenantFrom = location?.state?.tenantFrom || stripes.okapi.tenant;
+
+    switchAffiliation(stripes, tenantFrom, this.goBack);
   }
 
   render() {
-    if (this.isLoading()) {
-      return <PaneLoading defaultWidth="100%" />;
-    }
+    const {
+      stripes: { okapi },
+      location: { state },
+    } = this.props;
 
     return (
       <DataContext.Consumer>
         {data => (
           <ItemView
             {...this.props}
+            tenantTo={state?.tenantTo || okapi.tenant}
+            onCloseViewItem={this.onClose}
             referenceTables={data}
           />
         )}
