+++ src/views/ItemView.js
@@ -5,6 +5,7 @@ import {
   isEmpty,
   values,
   sortBy,
+  flowRight,
 } from 'lodash';
 import { parameterize } from 'inflected';
 
@@ -51,7 +52,10 @@ import {
   IntlConsumer,
   CalloutContext,
   checkIfUserInCentralTenant,
+  stripesConnect,
 } from '@folio/stripes/core';
+
+import { requestsStatusString } from '../Instance/ViewRequests/utils';
 
 import ModalContent from '../components/ModalContent';
 import { ItemAcquisition } from '../Item/ViewItem/ItemAcquisition';
@@ -86,11 +90,154 @@ import {
   WarningMessage,
   AdministrativeNoteList,
   ItemViewSubheader,
+  PaneLoading,
 } from '../components';
+
+const getRequestsPath = `circulation/requests?query=(itemId==:{itemid}) and status==(${requestsStatusString}) sortby requestDate desc&limit=1`;
 
 export const requestStatusFiltersString = map(REQUEST_OPEN_STATUSES, requestStatus => `requestStatus.${requestStatus}`).join(',');
 
 class ItemView extends React.Component {
+  static manifest = Object.freeze({
+    query: {},
+    itemsResource: {
+      type: 'okapi',
+      path: 'inventory/items/:{itemid}',
+      POST: { path: 'inventory/items' },
+      resourceShouldRefresh: true,
+      tenant: '!{tenantTo}',
+    },
+    markItemAsWithdrawn: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-withdrawn',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markItemAsMissing: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-missing',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsInProcess: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-in-process',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsInProcessNonRequestable: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-in-process-non-requestable',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsIntellectualItem: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-intellectual-item',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsLongMissing: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-long-missing',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsRestricted: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-restricted',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsUnavailable: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-unavailable',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    markAsUnknown: {
+      type: 'okapi',
+      POST: {
+        path: 'inventory/items/:{itemid}/mark-unknown',
+      },
+      clientGeneratePk: false,
+      fetch: false,
+    },
+    holdingsRecords: {
+      type: 'okapi',
+      path: 'holdings-storage/holdings/:{holdingsrecordid}',
+      tenant: '!{tenantTo}',
+    },
+    instanceRecords: {
+      type: 'okapi',
+      path: 'inventory/instances/:{id}',
+      resourceShouldRefresh: true,
+    },
+    servicePoints: {
+      type: 'okapi',
+      path: 'service-points',
+      records: 'servicepoints',
+      params: (_q, _p, _r, _l, props) => {
+        // Only one service point is of interest here: the SP used for the item's last check-in
+        // (if the item has a check-in). Iff that service point ID is found, add a query param
+        // to filter down to that one service point in the records returned.
+        const servicePointId = get(props.resources, 'itemsResource.records[0].lastCheckIn.servicePointId', '');
+        const query = servicePointId && `id==${servicePointId}`;
+        return query ? { query } : {};
+      },
+      resourceShouldRefresh: true,
+    },
+    staffMembers: {
+      type: 'okapi',
+      path: 'users',
+      records: 'users',
+      params: (_q, _p, _r, _l, props) => {
+        const staffMemberId = get(props.resources, 'itemsResource.records[0].lastCheckIn.staffMemberId', '');
+        const query = staffMemberId && `id==${staffMemberId}`;
+
+        return query ? { query } : null;
+      },
+      resourceShouldRefresh: true,
+    },
+    // return a count of the requests matching the given item and status
+    requests: {
+      type: 'okapi',
+      path: getRequestsPath,
+      records: 'requests',
+      PUT: { path: 'circulation/requests/%{requestOnItem.id}' },
+    },
+    openLoans: {
+      type: 'okapi',
+      path: 'circulation/loans',
+      params: {
+        query: 'status.name=="Open" and itemId==:{itemid}',
+      },
+      records: 'loans',
+    },
+    requestOnItem: {},
+    tagSettings: {
+      type: 'okapi',
+      records: 'configs',
+      path: 'configurations/entries?query=(module==TAGS and configName==tags_enabled)',
+    },
+  });
+
   static contextType = CalloutContext;
 
   constructor(props) {
@@ -452,7 +599,29 @@ class ItemView extends React.Component {
   getEntity = () => this.props.resources.itemsResource.records[0];
   getEntityTags = () => this.props.resources.itemsResource.records[0]?.tags?.tagList || [];
 
+  isLoading = () => {
+    const {
+      resources: {
+        instanceRecords,
+        itemsResource,
+        holdingsRecords,
+      },
+    } = this.props;
+
+    if (!itemsResource?.hasLoaded ||
+      !instanceRecords?.hasLoaded ||
+      !holdingsRecords?.hasLoaded) {
+      return true;
+    }
+
+    return false;
+  }
+
   render() {
+    if (this.isLoading()) {
+      return <PaneLoading defaultWidth="100%" />;
+    }
+
     const {
       resources: {
         itemsResource,
@@ -1536,15 +1705,24 @@ ItemView.propTypes = {
     })
   }).isRequired,
   resources: PropTypes.shape({
-    instanceRecords: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
+    instanceRecords: PropTypes.shape({
+      hasLoaded: PropTypes.bool,
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
     loanTypes: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
     requests: PropTypes.shape({
       records: PropTypes.arrayOf(PropTypes.object),
       other: PropTypes.object,
     }),
     loans: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
-    itemsResource: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
-    holdingsRecords: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
+    itemsResource: PropTypes.shape({
+      hasLoaded: PropTypes.bool,
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
+    holdingsRecords: PropTypes.shape({
+      hasLoaded: PropTypes.bool,
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
     callNumberTypes: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
     borrower: PropTypes.object,
     staffMembers: PropTypes.object,
@@ -1596,4 +1774,7 @@ ItemView.propTypes = {
   history: PropTypes.object.isRequired,
 };
 
-export default withLocation(ItemView);
+export default flowRight(
+  stripesConnect,
+  withLocation,
+)(ItemView);
